requirements.txt
fastapi==0.115.6
Pillow==11.0.0
psycopg2==2.9.10
Requests==2.32.3
torch==2.4.1+cu118
torch==2.5.1
torchvision==0.19.1+cu118
torchvision==0.20.1
tqdm==4.67.1
